listingautopilot/pyproject.toml
[tool.poetry]
name = "listingautopilot"
version = "0.1.0"
description = "AI service that converts product photos into Amazon-ready creative packs."
authors = ["Mithlesh Upadhyay <mithleshupadhyay515@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "listingautopilot", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.7"
python-dotenv = "^1.0"
openai = "^1.40"
google-generativeai = "^0.7"
anthropic = "^0.34"
streamlit = "^1.36"
pillow = "^10.4"
requests = "^2.32"
fastapi = "^0.111"
uvicorn = "^0.30"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0"
alembic = "^1.13"
psycopg2-binary = "^2.9"
prometheus-fastapi-instrumentator = "^7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
ruff = "^0.6"
mypy = "^1.11"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
